Change to decimal to aid currency comparisons; additional tests

+++ CheckoutServices/IScanService.cs
@@ -4,6 +4,6 @@
     {
         void Scan(string item);
 
-        double Total();
+        decimal Total();
     }
 }
